Fix for CheckIntegrity class exception handler
CheckIntegrity thread class will stop checking the remaining tables
in the list when it finds an error in COPY ... TO command. The
except clause was creating a new connection and closing the same
rightaway. Instead it should close the old connection and open a
new one (limitation by pygresql wrapping everything in a tnx).

This patch fixes the mentioned bug, also reported in #3

Signed-off-by: Aitor Cedres <acedres@pivotal.io>