PANIC on the segment kills the thread
When we encounter a PANIC in any of the queries performed during
the integrity exam, the thread for the affected dbid dies. This patch
will make sure we catch all types of exceptions when we query the
database tables. This way the thread will not die if we meet a PANIC
error in the segment.

Logging in spawn_threads has been slightly modified to always print
the segment and database.

Signed-off-by: Ignacio Elizaga <ielizaga@pivotal.io>