Add an option to set the number of threads spawned
This commit provides a command line option to limit the number of
threads spawned. It has a default value of 32.

This commit also provides the changes to CheckIntegrity thread to
use a semaphore before connecting to the DB. The semaphore capacity
will be set as the value specified in the command line option or
32 if none provided or invalid (0 or negative are invalid).

Signed-off-by: Aitor Cedres <acedres@pivotal.io>