Feature to support multiple schemas
This patch provides the functionality to allow multiple schema
check. It provides a new option to gpcheckintegrity to specify
a list of schemas. If such option is provided, only the tables
in the schema list will be checked.

Added a support function to check if a schema exists. If any
of the schemas in the list does not exists, it logs a warning
and continues.

This closes #12

Signed-off-by: Aitor Cedres <acedres@pivotal.io>