Fix CheckIntegrity thread exit on unexpected error
CheckIntegrity thread will report the current table as bad if the
thread hits an unexpected error. We only accept tables as corrupt if
 COPY table to dev-null returns an error. Unexpected error in the
thread should not report a table as corrupt. In this commit we add
a second except clause to CheckIntegrity.run() to differentaite
between unexpected errors/exceptions and errors from COPY to dev-null
 query.

Signed-off-by: Aitor Cedres <acedres@pivotal.io>