Adapting connect() function to return None
Our wrapper for connect() will exit (exit(1)) if the connection to
GPDB fails. This behavior is undesired within a thread function
because it might leave semaphores acquired. In this commit we change
connect() to return None if the connection fails. It is now
responsibility of the caller to check wether connect() returns None
or a connection object.

In the thread run() function, it will release the semaphore and
return. In case of facing a connect error during a pg.InternalError,
it will release the semaphore and return. If run() hits an unexpected
error (i.e. DB PANICs), it will log an error and break the loop.

Any error while trying to get information from master (get table
names, tables in a schema, segment configuration, etc) will finish
the program.

This one is related to #38

Signed-off-by: Aitor Cedres <acedres@pivotal.io>